blog/requirements.txt
Django==1.7.1
Markdown==2.5.1
Pillow==2.6.1
Pygments==2.0.1
Twisted==13.2.0
Unidecode==0.04.16
Werkzeug==0.9.6
altgraph==0.10.2
backports.ssl_match_hostname==3.4.0.2
bdist_mpkg==0.5.0
certifi==14.05.14
dictshield==0.4.4
django-uuslug==1.0.3
google-measurement-protocol==0.1.3
greenlet==0.4.5
image==1.3.4
jsonrpclib==0.1.3
macholib==1.5.1
markdown2==2.3.0
motor==0.3.4
pyjade==3.0.0
pymongo==2.7.1
python-dateutil==1.5
python-slugify==0.1.0
pytz==2013.7
requests==2.4.3
schematics==1.0-0
sockjs-tornado==1.0.1
tornado==4.0.2
tornadorpc==0.1.1
xlrd==0.9.3

